# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authorization Service for Medods",
        "title": "Medods Auth Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8095",
    "paths": {
...
        "/api/auth/refresh": {
            "post": {
                "description": "Refresh tokens (access and refresh token)",
                "tags": [
                    "auth"
                ],
                "summary": "Refresh tokens",
                "operationId": "refresh",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Refresh token",
                        "name": "refresh_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Refreshed Tokens",
                        "schema": {
                            "$ref": "#/definitions/model.TokenPair"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/status.Status"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "status.Status": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}